SEM3/ACNS/Assignment2/MedAPI/api/requirements.txt
flask
flask-cors
pillow
numpy
# Uncomment the following lines to use actual Huggingface models
transformers
torch
torchvision